Support for early-init.el
Some configuration frameworks for Emacs like Spacemacs, Doom, etc. have started
to make use of early-init.el which will be introduced by the pending release of
Emacs 27. This commit adds support for delegating the load of early-init.el to
the corresponding Chemacs profile.

It was necessary to alter the installation procedure for Chemacs. The README
features the new installation instructions.

Backwards compatibility is maintained: if Chemacs has been installed using the
old instructions it will continue to work, it merely won't be able to delegate
the load of early-init.el. That also means it is unnecessary for users to
re-install Chemacs using the new installation instructions unless they actually
use Emacs 27.