Set user-emacs-directory and server-name before early-init.el
Recently Spacemacs starting using the variable user-emacs-directory in the early
init file. This causes an error because when chemacs invokes the early init
file, the variable is not yet updated to the profile-specific value.

To fix this, this change sets the user-emacs-directory and server-name variables
before invoking the early init file.

It is intentional that the variables are set twice, just before invoking the
early init file, and again just before invoking the init file. This is done to
support Emacs versions older than 27.1, which will not enter the early init
logic during startup.